Moved import_metatdata.test to correct directory
